+++ src/process/process_read_instruction.c
@@ -10,7 +10,7 @@
 
 static int set_wait(process_t *process, int code)
 {
-    for (size_t i = 0; op_tab[i].mnemonique; i++) {
+    for (int64_t i = 0; op_tab[i].mnemonique; i++) {
         if (op_tab[i].code != code)
             continue;
         process->cycles_to_wait = op_tab[i].nbr_cycles;
@@ -21,7 +21,7 @@ static int set_wait(process_t *process, int code)
 
 int process_read_instruction(process_t *process, arena_t *arena)
 {
-    for (size_t i = 0; instructions[i].f; i++) {
+    for (int64_t i = 0; instructions[i].f; i++) {
         if (arena->arena[process->PC] == instructions[i].code) {
             instructions[i].f(process, arena);
             set_wait(process, instructions[i].code);
